fix(score): preserve scrubbed diagnostic file paths
Send useful file locations to the score API while redacting home-directory identities and embedded secrets.
