+++ packages/core/src/index.ts
@@ -111,6 +111,7 @@ export * from "./utils/resolve-github-actions-score-metadata.js";
 export * from "./utils/resolve-lint-batch-ordering.js";
 export * from "./utils/resolve-react-doctor-cache-dir.js";
 export * from "./utils/resolve-scan-concurrency.js";
+export * from "./utils/scrub-sensitive-paths.js";
 export * from "./utils/sort-diagnostics-stable.js";
 export * from "./utils/to-relative-path.js";
 export * from "./utils/warn-config-issue.js";
